Membuat aturan view,edit,delete data user agar user tidak dapat melakukan CRUD pada data user lainnya

+++ app/Http/Controllers/ChartController.php
@@ -12,7 +12,7 @@ class ChartController extends Controller
      */
     public function index()
     {
-        $charts = Chart::all();
+        $charts = Chart::where('user_id', auth()->id())->get();
         return view('charts.index', ['charts' => $charts]);
     }
 
@@ -31,10 +31,12 @@ class ChartController extends Controller
     {
         Chart::create([
             'kode_barang'=>$request->kode_barang,
+            'user_id'=>auth()->id(),
             'nama_barang'=>$request->nama_barang,
             'qty'=>$request->qty,
             'harga_barang'=>$request->harga_barang,
-            'total_harga_barang'=>$request->qty* $request->harga_barang
+            'total_harga_barang'=>$request->qty* $request->harga_barang,
+            
         ]);
 
         return redirect()->route('charts.index');
@@ -53,6 +55,12 @@ class ChartController extends Controller
      */
     public function edit(string $id)
     {
+        $uservalidation = Chart::where('id', $id)->where('user_id', auth()->id())->first();
+
+        if(!$uservalidation){
+            return abort(403, 'Anda tidak memiliki izin untuk mengedit barang ini.');
+        }
+
         $chart = Chart::find($id);
         return view('charts.edit', ['chart'=>$chart]);
     }
@@ -62,6 +70,13 @@ class ChartController extends Controller
      */
     public function update(Request $request, string $id)
     {
+
+        $uservalidation = Chart::where('id', $id)->where('user_id', auth()->id())->first();
+
+        if(!$uservalidation){
+            return abort(403, 'Anda tidak memiliki izin untuk mengedit barang ini.');
+        }
+
         Chart::find($id)->update([
             'kode_barang'=>$request->kode_barang,
             'nama_barang'=>$request->nama_barang,
@@ -78,6 +93,13 @@ class ChartController extends Controller
      */
     public function destroy(string $id)
     {
+
+        $uservalidation = Chart::where('id', $id)->where('user_id', auth()->id())->first();
+
+        if(!$uservalidation){
+            return abort(403, 'Anda tidak memiliki izin untuk menghapus barang ini.');
+        }
+
         Chart::find($id)->delete();
         return redirect()->route('charts.index');
     }
